+++ README.md
@@ -5,15 +5,19 @@ The App does following:
     2.Retrieve a customer with their linked order(s)
     3.Load all customers and their linked orders
 
-To build the App: mvn clean install
+To build the App: `mvn clean install`
 
 
 
-To Build the App without unit Tests: mvn clean install -DskipTests=true
+To Build the App without unit Tests: `mvn clean install -DskipTests=true`
 
 
-To start the App: jar -jar target/anycompany-0.0.1-SNAPSHOT.jar OR 
+To start the App: `java -jar target/anycompany-0.0.1-SNAPSHOT.jar `OR 
 
 
-mvn spring-boot:run
+`mvn spring-boot:run`
+
+Application is running on : `http://localhost:8080` 
+
+Access H2 Console on: `http://localhost:8080/h2-console/`
