+++ src/main/java/za.co.anycompany/model/Customer.java
@@ -2,13 +2,11 @@ package za.co.anycompany.model;
 
 import lombok.Getter;
 import lombok.Setter;
-import lombok.ToString;
-
 import javax.persistence.*;
 import java.io.Serializable;
 import java.util.Date;
 import java.util.List;
-import java.util.Set;
+
 
 @Getter
 @Setter
